Fix issue with foreign key column types

+++ api/database/migrations/2020_03_15_220126_create_cart_items_table.php
@@ -15,8 +15,8 @@ class CreateCartItemsTable extends Migration
     {
         Schema::create('cart_items', function (Blueprint $table) {
             $table->id();
-            $table->integer('cart_id');
-            $table->integer('product_id');
+            $table->unsignedBigInteger('cart_id');
+            $table->unsignedBigInteger('product_id');
             $table->integer('qty')->default(0);
             $table->float('line_total')->default(0);
